+++ src/tests/conftest.py
@@ -11,6 +11,10 @@ from normcap.ocr.models import OcrResult, TessArgs
 from normcap.utils import create_argparser
 
 
+def pytest_configure(config):
+    config.addinivalue_line("markers", "skip_on_gh: do not run during CI/CD on github")
+
+
 @pytest.fixture(scope="session")
 def capture() -> Generator[Capture, None, None]:
     """Create argparser and provide its default values."""